MYSQL/etl/requirements.txt
mysql-connector-python>=8.2.0
pandas>=2.1.0
python-dotenv>=1.0.0
pyodbc>=5.0.0
numpy>=1.24.0
requests>=2.31.0
